server/Server_1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
paho-mqtt==1.6.1
python-multipart==0.0.6
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.2 